Adds failure.rs error handling throughout.
Implements a `TibrvError` type, and an inner `ErrorKind` enum containing
variants likely to be encountered during use.

Introduces an `and_then` combinator for easily chaining onto the result
of a function returning a tibrv_status from the C library.

`TibrvError` derives `Fail`, so it can be chained in future with any
other libraries that use the Failure crate.