Migration for Tokio Reform
Changes to Tokio for 0.1 impose a couple of new constraints on futures
to be executed on the event loop.

All Futures now need to have a `'static` lifetime, and the new
tokio::run() and tokio::spawn() methods must return a Future with
Type<Item = ()>, so any top level Future needs mapped to a Future
producing ().