Bump bindgen version and change enum handling
Older versions of Bindgen fail to build with the latest stable Rust
compiler due to their dependency on old versions of syntex_syntax.

Updating to the latest version required a change to the way enums are
handled in the tibrv-sys crate:

Defining an enum binding from C previously generated code like this

[repr(C)]
enum tibrv_status {
    TIBRV_OK = 0,
    ...
}

The issue arose when the C library could return a value not included in
the Rust enum definition, in which case any match statement would no
longer be exhaustive. This could lead to undefined behaviour.

One solution here is to generate enums with the #[non_exhaustive]
annotation, however this would pin us to Rustc versions above 1.39 which
probably isn't acceptable for enterprise users.

As a compromise, we change how the enums are generated by bindgen,
creating instead a newtype wrapper and a list of associated constants,
matching on those instead.